Move processor connections  to `main.ts`
